PUT product

+++ src/main/java/com/SpringProfessional/dscommerce/repositories/ProductRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.SpringProfessional.dscommerce.entities.Product;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Integer >  {
+public interface ProductRepository extends JpaRepository<Product, Long >  {
 
 }
